Close enhance identification of IINA build, #4383
This commit will:
- Add Beta and Nightly Xcode build configurations to the Xcode project
- Add Beta.xcconfig and Nightly.xcconfig build configuration files for
  the new build configuration
- Add two new icons, AppIconBeta and AppIconNightly
- Use ASSETCATALOG_COMPILER_APPICON_NAME to select the correct
  application icon
- Add a key com.colliderli.iina.build.configuration to Info.plist to
  allow IINA to determine at runtime the build configuration used to
  generate the executable
- Add a property buildConfiguration to InfoDictionary that exposes the
  raw value in Info.plist
- Add a BuildType enumeration that corresponds to the build
  configurations in the Xcode project
- Add buildType, buildTypeIdentifier and isRelease properties to
  InfoDictionary to make it easy for other code to access the build
  configuration

This commit is only a start on the changes needed to implement #4383.
This commit only contains the changes to Xcode needed to use different
application icons for Beta and Nightly builds and to allow IINA to know
what type of build it is running. The new application icons have been
populated with document icons for testing. These must be replaced with
the appropriate icons once they have been generated. The commit does not
contain any changes to the welcome window or about window. Those changes
still have to be implemented.